+++ HISTORY.md
@@ -1,3 +1,9 @@
+## [v1.2.5]
+* added 'infos complementaires' to Tiers instance methods
+
+## (v1.2.4)
+* added SEARCHABLE mailing list
+
 ## [v1.2.0]
 * gemified
 * uses environment constant **J2R-DATA**